+++ trunk/Firmware/rtos_pilot/quaternion.h
@@ -1,11 +1,11 @@
-void quaternion_from_attitude (const float pitch, const float roll, const float yaw, float* q);
+void quaternion_from_attitude (const double pitch, const double roll, const double yaw, double* q);
 
-void quaternion_update_with_rates (const float rollrate, const float pitchrate, const float headingrate, float *q, const float dt);
+void quaternion_update_with_rates (const double rollrate, const double pitchrate, const double headingrate, double *q, const double dt);
 
-float quaternion_to_roll (const float* q);	
+double quaternion_to_roll (const double* q);	
 
-float quaternion_to_pitch(const float* q);
+double quaternion_to_pitch(const double* q);
 
-float quaternion_to_yaw(const float* q);
+double quaternion_to_yaw(const double* q);
 
-void quaternion_normalize(float* q);
+void quaternion_normalize(double* q);
